Change `CREATE` API to handle max_memory_sample_size and add a default evictor
Earlier the `CREATE` command just used to take one nameless argument,
which would have caused problems in future. Now the `CREATE` the evictor
is an optional argument, if nothing is passed by default `NOOP` evictor is used.
`CREATE` now also accepts an optional argument `SS` which is the max memory sample size
that the evictors use when determining an eviction candidate.